requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
pillow==10.2.0
pydantic==2.7.1
pydantic_core==2.18.2
sqlparse==0.4.4
typing_extensions==4.11.0
